Combine Data Streams
Implemented two different data obtaining methods
Get it all at once and get it when in need methods.
Get it all code is commented out in product service
in favour of get it when in need.

Handle ancillary streams.

Combined all the streams so we can use one async pipeline
instead of multiple pipelines. We use two pipelines one for
all the streams and separate one for error.

We separate error because combineLatest will wait for each stream to
emit at least one value. If we combined it with vm$ stream it will not
complete unless we have at least one error.