Add communication mechanism child to parent.
We can't use ngOnChanges because it only works with Input.
Therefore, we added EventEmitter decorated with @Output to
announce the changes to the parent.
EventEmitter will emit a payload of string, in this case out listFilter value.

The flow of the comminication as follows.

User add string on filter text field.
Critea template use 2 way binding to and our setter listFilter
is informed about the change.
Setter will emit a payload of the string value.

Parent template use event binding for the emitter and it will receieved the payload
and pass it to onValue change component.
Parent component use this method to call perform filter.