Add effect.
External factors such as servers can have side effects.
We need to keep our components pure from these side effects.
Therefore, component is not a suitable place to handle that.
We can't put that to reducers eithers since reducers are pure functions.
Effect is the answer for that.

In this, we create and use effect.
We add async pipe so that we don't need to manually subscribe.
We add error handling logic.