Use effect to perform update operation.
Instead of using product service directly on product-edit component,
we now dispatch and action to effect.

This effect will perform the operation using product service and dispatch an
action based on success or fail to reducer.

Reducer will create new state based on these success or fail action and update the store.

The create and delete will be implemented on next commit.

The reason for creating effect is it's design to handle operations with side effect.